refactor: move call-actor to a separate file and Actor related logic to utils
